build: Reduced runtime dependencies and added build of static library.
 * Added build of liblightnvm as a static library

 * Tests -- Changed linking from shared to static for
   - libcunit
   - liblightnvm

 * Examples -- Changed linking from shared to static for
   - liblightnvm

Providing a static version of liblightnvm allows building programs that need not
carry a runtime dependency on liblightnvm. That is, examples and tests
can be distributed without liblightnvm.[so|h].

Unfortunately then libudev, probably for good reasons, does not exists as a
static library. The runtime dependency on libudev by liblightnvm therefor
carries on.

However, with these changes less runtime dependencies are required allowing
for easier binary distribution.