Added structure encapsulating IO-subsystem results
liblightnvm provides the standard libc means of communicating errors, that is,
nvm_addr_* functions return -1 and sets errno. To provide a means to obtain
low-level result and status information the NVM_RET is introduced.

Pass it to nvm_addr_* functions and it will be filled out with low-level
return codes.