nvm_be_ioctl: add ioctl wrappers
Add `ioctl_wrap` function that wrap the NVMe ioctl calls and copies any
return values into the given `nvm_ret` (if not NULL).

This also adds special handling of the ioctl returning EINVAL, in which
case the NVMe "Invalid Field in Command" error is set in the given
`nvm_ret`.

With this patch all ioctl calls are now wrapped except the special ioctl
call that returns the namespace id. That call cannot be integrated with
the wrappers due to a different calling convention.

Also, update some documentation regarding `nvm_ret` and the backends.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>