cmake: use local/vendored spdk/dpdk
This changes the build process such that one does not need a locally
installed version of spdk/dpdk. Instead, it is cloned inside the project
and built there. Note that DPDK does not need to be built explicitly
anymore.

    git clone https://github.com/spdk/spdk.git
    git checkout v19.04
    git submodule update --init
    ./configure
    make

The build process of liblightnvm will then pick up needed headers and
libraries from there.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>